Fix some bugs in tensor.sml
Corrected some issues with the operation of map in TensorSlice, and fromto' in Range.  Things now _appear_ to be working correctly, so far as I can tell.

Furthermore, introduced some new functions for TensorSlice, namely `toTensor`, `sliceOutNewTensor` and `sliceOutNewTensor'`, which respectively take a pre-existing slice and turn it into a new tensor, take a current tensor and create a new tensor based on the specified inclusive index range, and take a current tensor and create a new tensor based on the specific exclusive index range.  Note that the latter means that if you want to slice out a single row or column, you need to specify a row or column index _one higher_ for the upper index than you want to slice out.